core/requirements.txt
asgi-redis==1.4.3
asgiref==1.1.2
attrs==17.2.0
autobahn==17.9.2
Automat==0.6.0
channels==1.1.8
constantly==15.1.0
daphne==1.3.0
dj-database-url==0.4.0
Django==1.11.5
django-cors-headers==2.1.0
django-filter==1.0.4
django-material==1.0.0
djangorestframework==3.6.4
drf-nested-routers==0.90.0
haikunator==1.0.1
hyperlink==17.3.1
incremental==17.5.0
msgpack-python==0.4.8
pytz==2017.2
redis==2.10.6
six==1.11.0
Twisted==17.5.0
txaio==2.8.2
zope.interface==4.4.2
